+++ src/tests/specs/mission-control/controllers/mission-control.controllers.satellite.spec.ts
@@ -1,0 +1,70 @@
+import {
+	idlFactoryMissionControl,
+	idlFactoryOrbiter,
+	idlFactorySatellite,
+	type MissionControlActor,
+	type OrbiterActor,
+	type SatelliteActor
+} from '$declarations';
+import { PocketIc, type Actor } from '@dfinity/pic';
+import { Ed25519KeyIdentity } from '@icp-sdk/core/identity';
+import type { Principal } from '@icp-sdk/core/principal';
+import { JUNO_AUTH_ERROR_NOT_ADMIN_CONTROLLER } from '@junobuild/errors';
+import { inject } from 'vitest';
+import { ORBITER_CONTROLLER_ERR_MSG } from '../../../constants/orbiter-tests.constants';
+import { missionControlUserInitArgs } from '../../../utils/mission-control-tests.utils';
+import {
+	controllersInitArgs,
+	MISSION_CONTROL_WASM_PATH,
+	ORBITER_WASM_PATH,
+	SATELLITE_WASM_PATH
+} from '../../../utils/setup-tests.utils';
+
+describe('Mission Control > Controllers > Satellite', () => {
+	let pic: PocketIc;
+	let actor: Actor<MissionControlActor>;
+
+	let satelliteId: Principal;
+
+	const controller = Ed25519KeyIdentity.generate();
+
+	beforeAll(async () => {
+		pic = await PocketIc.create(inject('PIC_URL'));
+
+		const userInitArgs = (): Uint8Array => missionControlUserInitArgs(controller.getPrincipal());
+
+		const { actor: c, canisterId: missionControlId } = await pic.setupCanister<MissionControlActor>(
+			{
+				idlFactory: idlFactoryMissionControl,
+				wasm: MISSION_CONTROL_WASM_PATH,
+				arg: userInitArgs(),
+				sender: controller.getPrincipal()
+			}
+		);
+
+		actor = c;
+
+		const { canisterId: cId } = await pic.setupCanister<SatelliteActor>({
+			idlFactory: idlFactorySatellite,
+			wasm: SATELLITE_WASM_PATH,
+			arg: controllersInitArgs([controller.getPrincipal(), missionControlId]),
+			sender: controller.getPrincipal()
+		});
+
+		satelliteId = cId;
+
+		actor.setIdentity(controller);
+	});
+
+	afterAll(async () => {
+		await pic?.tearDown();
+	});
+
+	it('should not be able to set a satellite because mission control is effectively not a controller', async () => {
+		const { set_satellite } = actor;
+
+		await expect(set_satellite(satelliteId, ['Hello'])).rejects.toThrowError(
+			new RegExp(JUNO_AUTH_ERROR_NOT_ADMIN_CONTROLLER)
+		);
+	});
+});
